Solutions/ai-bartender-agent/aws/services/image-processing/src/processImage/requirements.txt
aws-lambda-powertools[tracer]>=2.0.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
Pillow>=10.2.0
